Update function comments, add radius function

+++ src/math_helpers.h
@@ -14,7 +14,7 @@ struct coordPlanet {
 	bool westIsLeft; // if west is on the right side of image (after rotation) -> mirrored
 };
 
-deltacoords rayCenter(coordinates approximateCenter, image* frame, int32_t numberRays, config* cfg);
+deltacoords rayCenter(coordinates approximateCenter, image* frame, int16_t numberRays, config* cfg);
 deltacoords massCenter(image* frame, config* cfg);
 
 // fit a circle to a list of coordinates with least squares
